+++ util/src/test/java/net/automatalib/util/automata/fsa/NFAsTest.java
@@ -15,17 +15,99 @@
  */
 package net.automatalib.util.automata.fsa;
 
+import net.automatalib.automata.fsa.NFA;
 import net.automatalib.automata.fsa.impl.compact.CompactDFA;
 import net.automatalib.automata.fsa.impl.compact.CompactNFA;
+import net.automatalib.util.automata.Automata;
 import net.automatalib.words.Alphabet;
 import net.automatalib.words.Word;
 import net.automatalib.words.impl.Alphabets;
 import org.testng.Assert;
 import org.testng.annotations.Test;
 
-@Test
 public class NFAsTest {
 
+    private static final boolean[] VECTOR_1 = {true, true, false, false};
+    private static final boolean[] VECTOR_2 = {true, false, true, false};
+
+    // The precomputed results of applying the operations on VECTOR_1 and VECTOR_2
+    private static final boolean[] AND_RESULT = {true, false, false, false};
+    private static final boolean[] OR_RESULT = {true, true, true, false};
+    private static final boolean[] XOR_RESULT = {false, true, true, false};
+    private static final boolean[] EQUIV_RESULT = {true, false, false, true};
+    private static final boolean[] IMPL_RESULT = {true, false, true, true};
+
+    private final Alphabet<Integer> testAlphabet;
+    private final CompactNFA<Integer> testNfa1, testNfa2;
+
+    public NFAsTest() {
+        this.testAlphabet = Alphabets.integers(0, 0);
+        this.testNfa1 = forVector(VECTOR_1);
+        this.testNfa2 = forVector(VECTOR_2);
+    }
+
+    private CompactNFA<Integer> forVector(boolean... boolVec) {
+        if (boolVec.length == 0) {
+            throw new IllegalArgumentException("Length of vector must be non-zero");
+        }
+
+        CompactNFA<Integer> result = new CompactNFA<>(testAlphabet, boolVec.length);
+
+        int first = result.addInitialState(boolVec[0]);
+        int prev = first;
+
+        for (int i = 1; i < boolVec.length; i++) {
+            int next = result.addState(boolVec[i]);
+            result.addTransition(prev, 0, next);
+            prev = next;
+        }
+
+        result.addTransition(prev, 0, first);
+
+        return result;
+    }
+
+    @Test
+    public void testAnd() {
+        NFA<?, Integer> expected = forVector(AND_RESULT);
+        NFA<?, Integer> actual = NFAs.and(testNfa1, testNfa2, testAlphabet);
+
+        assertEquivalence(actual, expected, testAlphabet);
+    }
+
+    @Test
+    public void testOr() {
+        NFA<?, Integer> expected = forVector(OR_RESULT);
+        NFA<?, Integer> actual = NFAs.or(testNfa1, testNfa2, testAlphabet);
+
+        assertEquivalence(actual, expected, testAlphabet);
+    }
+
+    @Test
+    public void testXor() {
+        NFA<?, Integer> expected = forVector(XOR_RESULT);
+        NFA<?, Integer> actual = NFAs.xor(testNfa1, testNfa2, testAlphabet);
+
+        assertEquivalence(actual, expected, testAlphabet);
+    }
+
+    @Test
+    public void testEquiv() {
+        NFA<?, Integer> expected = forVector(EQUIV_RESULT);
+        NFA<?, Integer> actual = NFAs.equiv(testNfa1, testNfa2, testAlphabet);
+
+        assertEquivalence(actual, expected, testAlphabet);
+    }
+
+    @Test
+    public void testImpl() {
+        NFA<?, Integer> expected = forVector(IMPL_RESULT);
+        NFA<?, Integer> actual = NFAs.impl(testNfa1, testNfa2, testAlphabet);
+
+        assertEquivalence(actual, expected, testAlphabet);
+    }
+
+    @Test
     public void testDeterminize() {
         Alphabet<Integer> alphabet = Alphabets.integers(0, 1);
 
@@ -44,5 +126,12 @@ public class NFAsTest {
         CompactDFA<Integer> dfa = NFAs.determinize(nfa);
 
         Assert.assertEquals(dfa.size(), 2);
+    }
+
+    private <I> void assertEquivalence(NFA<?, I> nfa1, NFA<?, I> nfa2, Alphabet<I> inputs) {
+        Assert.assertTrue(Automata.testEquivalence(NFAs.determinize(nfa1, inputs),
+                                                   NFAs.determinize(nfa2, inputs),
+                                                   inputs));
+
     }
 }